Moved the models to their own abstraction
