--- .travis.yml
@@ -1,7 +1,0 @@
-language: node_js
-node_js:
- - "12"
-install:
- - npm i
-script:
- - npm run test
