+++ src/models/access.ts
@@ -1,0 +1,9 @@
+export interface access {
+  id: number,
+  role_id: number,
+  route: string,
+  create: boolean,
+  read: boolean,
+  update: boolean,
+  delete: boolean
+}
