+++ src/models/owner.ts
@@ -1,0 +1,5 @@
+export interface owner {
+  id: number,
+  name: string,
+  active: boolean
+}
