+++ src/models/project.ts
@@ -1,0 +1,8 @@
+export interface project {
+  id: number
+  owner_id: number
+  name: string
+  description: string
+  active: boolean
+  budget: number
+}
