+++ src/models/projectDetail.ts
@@ -1,0 +1,11 @@
+export interface projectDetail {
+  id: number
+  time_id: number
+  name: string
+  description: string
+  user: string
+  hours: number
+  project_hours: number
+  percent: number
+  color?: string
+}
