+++ src/models/role.ts
@@ -1,0 +1,6 @@
+export interface role {
+  id: number,
+  name: string,
+  active: boolean,
+  color: string
+}
