+++ src/models/type.ts
@@ -1,0 +1,5 @@
+export interface type {
+  id: number,
+  name: string,
+  active: boolean
+}
