+++ src/models/user.ts
@@ -1,0 +1,8 @@
+export interface user {
+  username: string,
+  password: string,
+  id: number,
+  active: boolean,
+  role: string,
+  role_id: number
+}
