+++ src/routes/StatsRoutes.ts
@@ -1,11 +1,13 @@
 import { Router } from 'express'
 import { groupBy } from '../utils/json'
 import { Client } from 'pg'
-import ProjectsService, { projectDetail } from '../services/ProjectService'
-import OwnerService, { owner } from '../services/OwnerService'
+import ProjectsService from '../services/ProjectService'
+import OwnerService from '../services/OwnerService'
 import { sidebarComponent } from '../components/sidebar/sidebar'
 import { hasAccess } from '../utils/auth'
 import { RoleService } from '../services/RoleService'
+import { owner } from '../models/owner'
+import { projectDetail } from '../models/projectDetail'
 
 export function StatsRoutes (pgClient: Client): Router {
   const router = Router()
