+++ src/services/RoleService.ts
@@ -1,22 +1,7 @@
+import { access } from "../models/access"
+import { role } from "../models/role"
 import { sqlString } from "../utils/sqlStrings"
 import DbService from "./DbService"
-
-export interface role {
-  id: number,
-  name: string, 
-  active: boolean,
-  color: string
-}
-
-export interface access {
-  id: number,
-  role_id: number,
-  route: string,
-  create: boolean,
-  read: boolean,
-  update: boolean,
-  delete: boolean
-}
 
 export class RoleService extends DbService{
   updateRole(arg0: number, role: string, color: string): void {
@@ -45,7 +30,7 @@ export class RoleService extends DbService{
 
   async getAccessByRouteAndRole(route: string, roleId: number): Promise<access> {
     const query = `
-    select a.id, a.route, a."create", a.read, a.update, a.delete 
+    select a.id, a.route, a."create", a.read, a.update, a.delete
     from access a
     where role_id = ${roleId}
     and route = '${sqlString(route)}';`
@@ -54,7 +39,7 @@ export class RoleService extends DbService{
 
   async getAccessByRole(roleId: number): Promise<access[]> {
     const query = `
-    select a.id, a.route, a."create", a.read, a.update, a.delete 
+    select a.id, a.route, a."create", a.read, a.update, a.delete
     from role
     inner join access a on role.id = a.role_id
     where role.id = ${roleId};`
@@ -71,7 +56,7 @@ export class RoleService extends DbService{
 
   async updateAccess(accessId: number, route: string, create: boolean, read: boolean, update: boolean, _delete: boolean): Promise<number> {
     return (await this.client.query(`
-      update access 
+      update access
       set route = '${sqlString(route)}',
       "create" = ${create},
       read = ${read},
